feat: add strutucture project
